+++ frontend/ui/src/components/kanban-board/board/index.tsx
@@ -15,7 +15,6 @@ import {
   DropdownMenu 
 } from "@incmix/ui"
 import { 
-  Plus, 
   Loader2, 
   Settings, 
   MoreVertical,
@@ -24,7 +23,6 @@ import {
 import { isCardData, isCardDropTargetData, isColumnData, isDraggingACard, KanbanColumn, useKanban } from "@incmix/store"
 
 import { BoardColumn } from "./board-column"
-import { AddTaskForm } from "../shared/add-task-form"
 import { TaskCardDrawer } from "../shared/task-card-drawer"
 import { CreateColumnForm } from "./create-column-form"
