+++ frontend/ui/src/components/kanban-board/shared/task-card-components/index.ts
@@ -1,0 +1,22 @@
+// task-card-components/index.ts
+
+// Main components
+export { TaskDrawerSheet } from './task-drawer-sheet'
+export { TaskHeader } from './task-header'
+export { TaskActionButtons } from './task-action-buttons'
+export { TaskTitleSection } from './task-title-section'
+export { TaskDescriptionSection } from './task-description-section'
+export { TaskDatesSection } from './task-dates-section'
+export { TaskTagsSection } from './task-tags-section'
+export { TaskAssigneesSection } from './task-assignees-section'
+export { TaskSubtasksSection } from './task-subtasks-section'
+export { TaskCommentsSection } from './task-comments-section'
+export { SubtaskItem } from './subtask-item'
+
+// Hooks
+export { useTaskActions } from './hooks/use-task-actions'
+export { useTaskDrawerState } from './hooks/use-task-drawer-state'
+
+// Utils and types
+export * from './utils/types'
+export * from './utils/task-utils'
